Unknown determining byte fix
When a full RTP packet could not be extracted in the same
callback, `onData` failed to determine what type of packet
that's currently being handled because `onInterleaveData`
extracted the determining byte from the ByteArray.

To fix this, only peek the data until we know we have enough
to extract a full RTP packet.