RTMPClient: Honor port
If a port is specified, connect to that port. If port is not specified,
it's important to leave it out of the URL (e.g. not to use the default 1935).
If the port is left out, NetStream will attempt different ways to reach the
server, both by RTMP and RTMPT on different ports.