Fix start playing even if previous socket is invalid
If the previous socket got invalid the teardown request will throw an
error. Since the play function will stop the current client and wait for
the STOPPED event to trigger the new url will never be played.

This fix makes sure the STOPPED event gets triggered even if the socket
is invalid.