Add frame by frame mode
Frame by frame mode lets the user choose when to play each chunk of
data. When a chunk becomes ready an event is fired, to play it the user
calls playFrames with the timestamp until which frames should be added
to the play buffer.

If the data connection is closed before the stream is finished playing
but after all data is received, dont stop the player until flash has
finished playing.