Don't force a close on POST channel after 2s
Most browsers only allow a specific number of concurrent open requests.
If mulitple locomote players are used, that number might be reached.
Keeping the POST channel open unnecessary long might impact the time it
takes to setup the RTSP session.
Also, when forcing a close on the POST channel after a given time the
request might never have been sent. Instead the POST channel is only
closed once the next request is about to be sent.