Circumvent IE URLStream open event defect
IE/Edge will emit the open event for the URLStream load
prematurely.
This will have the effect that the POST with the OPTIONS
request will arraive before the GET channel is set up.
Added a check if IE is used, then give some slack time
for the GET channel to be established before emitting
the 'connected' Event.